Moved test for avatarLogic to different file
